requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.1
python-dotenv==1.0.1
openpyxl==3.1.5

# PostgreSQL driver — required in production, not needed for SQLite locally
psycopg2-binary==2.9.9
